pyproject.toml
[project]
name = "torch-webgpu"
version = "0.0.1"
description = " WebGPU backend for PyTorch"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Jędrzej Maczan", email = "jedrzejpawel@maczan.pl" },
]
dependencies = [
    "python-dotenv",
    "torch==2.9.0",
    "torchvision==0.24",
    "numpy"
]

[build-system]
requires = ["setuptools", "wheel", "torch"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "python"}

[tool.setuptools.packages.find]
where = ["python"]